+++ src/Command/SearchImportCommand.php
@@ -46,7 +46,7 @@ class SearchImportCommand extends IndexCommand
 
                 $output->writeln(sprintf(
                     'Indexed <comment>%s / %s</comment> %s entities into %s index',
-                    $response[$indexName],
+                    isset($response[$indexName]) ? $response[$indexName] : 0,
                     count($entities),
                     $entityClassName,
                     '<info>' . $config['prefix'] . $indexName . '</info>'
@@ -55,9 +55,10 @@ class SearchImportCommand extends IndexCommand
                 $page++;
                 $repository->clear();
             } while (count($entities) >= $config['batchSize']);
+
+            $repository->clear();
         }
 
-        $repository->clear();
 
         $output->writeln('<info>Done!</info>');
     }