+++ tests/config/default.php
@@ -8,10 +8,12 @@ return [
         'posts' => [
             'class' => 'Algolia\SearchBundle\Entity\Post',
             'enable_serializer_groups' => false,
+            'index_if' => null,
         ],
         'comments' => [
             'class' => 'Algolia\SearchBundle\Entity\Comment',
             'enable_serializer_groups' => false,
+            'index_if' => null,
         ],
     ]
 ];